Move switch function for use in python code
* Avoid import issues by providing a separate module for this function
* Document using feature flipping within Python
