+++ bedrock/mozorg/templatetags/misc.py
@@ -24,7 +24,6 @@ except ImportError:
 
 import bleach
 import jinja2
-from decouple import config
 from django_jinja import library
 
 from bedrock.base.urlresolvers import reverse
@@ -49,25 +48,6 @@ def add_string_to_image_url(url, addition):
 def convert_to_high_res(url):
     """Convert a file name to the high-resolution version."""
     return add_string_to_image_url(url, 'high-res')
-
-
-@library.global_function
-def switch(name):
-    """A template helper that replaces waffle
-
-    * All calls default to True when DEV setting is True.
-    * If the env var is explicitly false it will be false even when DEV = True.
-    * Otherwise the call is False by default and True is a specific env var exists and is truthy.
-
-    For example:
-
-        {% if switch('dude-and-walter') %}
-
-    would check for an environment variable called `SWITCH_DUDE_AND_WALTER`. The string from the
-    `switch()` call is converted to uppercase and dashes replaced with underscores.
-    """
-    env_name = 'SWITCH_' + name.upper().replace('-', '_')
-    return config(env_name, default=settings.DEV, cast=bool)
 
 
 @library.global_function
